move browse into entry
